organize scripts, add inventory system, add sword, add bow

+++ Assets/Scripts/Items/IItem.cs
@@ -2,7 +2,9 @@ using UnityEngine;
 
 public interface IItem
 {
-    public float Damage { get; set; }
+    public int Damage { get; set; }
     
     public void UseItem(Vector3 position, Vector2 direction);
+
+    
 }
